requirements.txt
python-docx
pandas
requests
openpyxl
tqdm
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles